requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.7
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.6.0
django-reversion==2.0.11
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
